RT #41995 Implementation of Pod::Simple::Select
I implemented the equivalent functionality of Pod::Select (now being discouraged) into Pod::Simple::Select
I added a "raw_mode" in Pod::Simple::Blackbox. This mode forces Pod::Simple to parse very lazily and report entire, raw Pod lines.
I also implemented a module called Pod::Simple::Select that is a subclass of Pod::Simple and uses this raw mode.
There is an object-oriented and functional interface. The functional interface to this module has a single function, podselect(), that is a drop-in replacement for the Pod::Select::podselect() function.